backend/requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.3
django-extra-fields==3.0.2
django-filter==23.1
django-import-export==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.2.0
django-debug-toolbar==3.2.4
et-xmlfile==1.1.0
executing==1.2.0
flake8==6.0.0
gunicorn==20.0.4
idna==3.4
ipython==8.14.0
jedi==0.19.0
MarkupPy==1.14
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.5.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
webcolors==1.13
xlrd==2.0.1
xlwt==1.3.0
reportlab==3.6.1
xhtml2pdf==0.2.4
django-cors-headers==3.13.0
django-colorfield==0.11.0